+++ examples/advanced_example/lib/tiles.dart
@@ -32,27 +32,17 @@ ScheduleTileComponents get scheduleTileComponents {
 abstract class BaseEventTile extends StatelessWidget {
   final Event event;
   final KalenderDateTimeRange tileRange;
-  const BaseEventTile({
-    super.key,
-    required this.event,
-    required this.tileRange,
-  });
+  const BaseEventTile({super.key, required this.event, required this.tileRange});
 
   Color get color => event.person.color;
-  Color textColor(Color color) =>
-      color.computeLuminance() > 0.5 ? Colors.black : Colors.white;
+  Color textColor(Color color) => color.computeLuminance() > 0.5 ? Colors.black : Colors.white;
 
   bool get continuesAfter => event.dateTimeRange.end.isAfter(tileRange.end);
-  bool get continuesBefore =>
-      event.dateTimeRange.start.isBefore(tileRange.start);
-  String title(BuildContext context) =>
-      "${event.person.toString()}: ${event.title} (${event.id})";
+  bool get continuesBefore => event.dateTimeRange.start.isBefore(tileRange.start);
+  String title(BuildContext context) => "${event.person.toString()}: ${event.title} (${event.id})";
 
   static BorderRadius defaultBorderRadius = BorderRadius.circular(8);
-  BoxDecoration get decoration => BoxDecoration(
-    color: color.withAlpha(150),
-    borderRadius: defaultBorderRadius,
-  );
+  BoxDecoration get decoration => BoxDecoration(color: color.withAlpha(150), borderRadius: defaultBorderRadius);
 }
 
 class EventTile extends BaseEventTile {
@@ -76,21 +66,12 @@ class EventTile extends BaseEventTile {
 }
 
 class MultiDayEventTile extends BaseEventTile {
-  const MultiDayEventTile({
-    super.key,
-    required super.event,
-    required super.tileRange,
-  });
-  static MultiDayEventTile builder(
-    BuildContext context,
-    CalendarEvent event,
-    KalenderDateTimeRange tileRange,
-  ) {
+  const MultiDayEventTile({super.key, required super.event, required super.tileRange});
+  static MultiDayEventTile builder(BuildContext context, CalendarEvent event, KalenderDateTimeRange tileRange) {
     return MultiDayEventTile(event: event as Event, tileRange: tileRange);
   }
 
-  EdgeInsets get padding =>
-      const EdgeInsets.symmetric(vertical: 1, horizontal: 4);
+  EdgeInsets get padding => const EdgeInsets.symmetric(vertical: 1, horizontal: 4);
 
   static Key getKey(String id) => Key('MultiDayEventTile-$id');
 
@@ -112,22 +93,13 @@ class MultiDayEventTile extends BaseEventTile {
 }
 
 class OverlayEventTile extends BaseEventTile {
-  const OverlayEventTile({
-    super.key,
-    required super.event,
-    required super.tileRange,
-  });
+  const OverlayEventTile({super.key, required super.event, required super.tileRange});
 
-  static OverlayEventTile builder(
-    BuildContext context,
-    CalendarEvent event,
-    KalenderDateTimeRange tileRange,
-  ) {
+  static OverlayEventTile builder(BuildContext context, CalendarEvent event, KalenderDateTimeRange tileRange) {
     return OverlayEventTile(event: event as Event, tileRange: tileRange);
   }
 
-  EdgeInsets get padding =>
-      const EdgeInsets.symmetric(vertical: 1, horizontal: 4);
+  EdgeInsets get padding => const EdgeInsets.symmetric(vertical: 1, horizontal: 4);
 
   @override
   Widget build(BuildContext context) {
@@ -137,29 +109,15 @@ class OverlayEventTile extends BaseEventTile {
           left: continuesBefore ? 20 : 0,
           right: continuesAfter ? 20 : 0,
           child: DecoratedBox(
-            decoration: BoxDecoration(
-              color: color.withAlpha(150),
-              borderRadius: BorderRadius.circular(8),
-            ),
+            decoration: BoxDecoration(color: color.withAlpha(150), borderRadius: BorderRadius.circular(8)),
             child: Padding(
               padding: padding,
-              child: Text(
-                title(context),
-                style: TextStyle(color: textColor(color)),
-              ),
+              child: Text(title(context), style: TextStyle(color: textColor(color))),
             ),
           ),
         ),
-        if (continuesAfter)
-          const Align(
-            alignment: Alignment.centerRight,
-            child: Icon(Icons.chevron_right),
-          ),
-        if (continuesBefore)
-          const Align(
-            alignment: Alignment.centerLeft,
-            child: Icon(Icons.chevron_left),
-          ),
+        if (continuesAfter) const Align(alignment: Alignment.centerRight, child: Icon(Icons.chevron_right)),
+        if (continuesBefore) const Align(alignment: Alignment.centerLeft, child: Icon(Icons.chevron_left)),
       ],
     );
   }
@@ -168,16 +126,9 @@ class OverlayEventTile extends BaseEventTile {
 class FeedbackTile extends StatelessWidget {
   final Event event;
   final Size dropTargetWidgetSize;
-  const FeedbackTile({
-    super.key,
-    required this.event,
-    required this.dropTargetWidgetSize,
-  });
+  const FeedbackTile({super.key, required this.event, required this.dropTargetWidgetSize});
   static FeedbackTile builder(BuildContext context, CalendarEvent event, Size dropTargetWidgetSize) {
-    return FeedbackTile(
-      event: event as Event,
-      dropTargetWidgetSize: dropTargetWidgetSize,
-    );
+    return FeedbackTile(event: event as Event, dropTargetWidgetSize: dropTargetWidgetSize);
   }
 
   @override