+++ examples/riverpod/lib/main.dart
@@ -28,7 +28,8 @@ final eventsProvider = Provider<EventsController>((ref) => DefaultEventsControll
 /// The view configurations the user can pick from.
 final viewConfigurationsProvider = Provider<List<ViewConfiguration>>((ref) {
   final now = DateTime.now();
-  final displayRange = KalenderDateTimeRange(start: now.copyWith(day: now.day - 365), end: now.copyWith(day: now.day + 365));
+  final displayRange =
+      KalenderDateTimeRange(start: now.copyWith(day: now.day - 365), end: now.copyWith(day: now.day + 365));
   return [
     MultiDayViewConfiguration.week(displayRange: displayRange, firstDayOfWeek: 1),
     MultiDayViewConfiguration.singleDay(displayRange: displayRange),