+++ test/widgets/resize_handle_style_test.dart
@@ -21,7 +21,7 @@ void main() {
     eventsController = DefaultEventsController();
     calendarController = CalendarController();
     eventId = eventsController.addEvent(
-      CalendarEvent(dateTimeRange: KalenderDateTimeRange(start: DateTime(2025, 1, 1, 1), end: DateTime(2025, 1, 1, 4))),
+      CalendarEvent(start: DateTime(2025, 1, 1, 1), end: DateTime(2025, 1, 1, 4)),
     );
   });